(cleanup) Move nav config into /config folder
