+++ app/components/Feed.js
@@ -22,7 +22,7 @@ import Menu from './Menu.js';
 import EventCard from './EventCard';
 import configURL from './../config/config.js';
 
-import _navigate from './navigateConfig.js';
+import _navigate from '../config/navigateConfig.js';
 
 
 export default class Feed extends Component {
